\name{AC15}
\docType{data}
\alias{AC15}
\title{Typometry dataset AC15}
\description{
  This data set gives the lengths, widths and thicknesses of some stone
  tools.
}
\usage{AC15}
\format{A data frame containing 28 observations of 3 variables.}
\source{Provided by Dr. Elisabetta Starnini.}

\keyword{datasets}
