\name{D7CAN2}
\docType{data}
\alias{D7CAN2}
\title{Typometry dataset D7CAN2}
\description{
  This data set gives the lengths, widths and thicknesses of some stone
  tools.
}
\usage{D7CAN2}
\format{A data frame containing 281 observations of 3 variables.}
\source{Provided by Dr. Elisabetta Starnini.}

\keyword{datasets}
