\name{tipom-package}
\alias{tipom-package}
\encoding{utf8}
\alias{tipom}
\docType{package}
\title{
Automated measure-based classification of prehistoric flint tools
}
\description{
TIPOM is based on a methodology that was developed in the 1960s by
Bernardino Bagolini. The basic idea is to use the three simple measures
of length, width and thickness of each flint artefact to classify them
in discrete groups, and assess their function based on such classification.
}
\details{
\tabular{ll}{
Package: \tab tipom\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-16\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Who wrote it
Stefano Costa <steko@iosa.it> with Luca Bianconi <lc.bianconi@googlemail.com>
Maintainer: Stefano Costa <steko@iosa.it>
}
\references{
B.Bagolini, “Ricerche sui manufatti litici preistorici non ritoccati”, Annali Università di Ferrara, n. s., sez XV, I, 10 (1968), pag. 195 sgg

S. Costa, L. Bianconi and E. Starnini, “” in F. Cantone (ed.), ARCHEOFOSS. Open Source, Free Software e Open Format nei processi di ricerca archeologica. Atti del VI Workshop (Napoli, 9-10 giugno 2011), Napoli:Naus 2012, pp. 211-218.
}

\keyword{ package }

